[test] add decryption tests
